backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
torch>=1.8.0
torchvision>=0.9.0
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
